refactor: move context if-else conditions into version_compat
