+++ providers/github/src/airflow/providers/github/version_compat.py
@@ -36,12 +36,16 @@ AIRFLOW_V_3_0_PLUS = get_base_airflow_version_tuple() >= (3, 0, 0)
 
 if AIRFLOW_V_3_0_PLUS:
     from airflow.sdk import BaseOperator, BaseSensorOperator
+    from airflow.sdk.definitions.context import Context
 else:
     from airflow.models import BaseOperator
     from airflow.sensors.base import BaseSensorOperator  # type: ignore[no-redef]
+    from airflow.utils.context import Context
+
 
 __all__ = [
     "AIRFLOW_V_3_0_PLUS",
     "BaseOperator",
     "BaseSensorOperator",
+    "Context",
 ]
